update both api and tests
