+++ mercury/tests/test_radio_receiver.py
@@ -39,28 +39,6 @@ class TestRadioReceiverView(TestCase):
         self.uuid = "d81cac8d-26e1-4983-a942-1922e54a943d"
         self.uuid2 = "d81cac8d-26e1-4983-a942-1922e54a943a"
 
-    def post_radio_data(self):
-        # POST sensor data to the radioreceiver url
-        response = self.client.post(
-            reverse(self.post_url, args=[self.uuid]),
-            data={
-                "sensor_id": 1,
-                "values": {"power": "2", "speed": 1},
-                "date": datetime.datetime(2020, 2, 2, 20, 21, 22),
-            },
-        )
-        return response
-
-    def post_defect_data(self):
-        response = self.client.post(
-            reverse(self.post_url, args=[self.uuid]),
-            data={
-                "values": {"power": "2", "speed": 1},
-                "date": datetime.datetime(2020, 2, 2, 20, 21, 22),
-            },
-        )
-        return response
-
     def test_Radio_Receiver_GET_No_Related_Event(self):
         response = self.client.get(reverse(self.get_url, args=[self.uuid2]))
         self.assertEqual(400, response.status_code)
@@ -144,27 +122,4 @@ class TestRadioReceiverView(TestCase):
                 "fake": 1,
             },
         )
-        self.assertEqual(200, response.status_code)
-
-    def test_Radio_Receiver_POST_Event_Not_Exist(self):
-        response = self.client.post(reverse(self.get_url, args=[self.uuid2]))
-        self.assertEqual(400, response.status_code)
-        self.assertEqual("Wrong uuid in url", json.loads(response.content))
-
-    @mock.patch("mercury.models.AGEvent.objects.get", fake_event)
-    def test_Radio_Receiver_POST_Missing_Params(self):
-        response = self.post_defect_data()
-        self.assertEqual(400, response.status_code)
-        self.assertTrue("Missing required params" in json.loads(response.content))
-
-    @mock.patch("mercury.models.AGEvent.objects.get", fake_event)
-    def test_Radio_Receiver_POST_Fail_to_Save(self):
-        response = self.post_radio_data()
-        self.assertEqual(400, response.status_code)
-
-    @mock.patch("mercury.models.AGEvent.objects.get", fake_event)
-    @mock.patch("mercury.serializers.AGMeasurementSerializer.is_valid", fake_valid)
-    @mock.patch("mercury.serializers.AGMeasurementSerializer.save", fake_valid)
-    def test_Radio_Receiver_POST_Event_Success(self):
-        response = self.post_radio_data()
         self.assertEqual(200, response.status_code)
