+++ mercury/views/radioreceiver.py
@@ -2,7 +2,6 @@ import datetime
 import glob
 import json
 import os
-import subprocess
 import sys
 
 import serial
@@ -10,9 +9,9 @@ from django.core.serializers.json import DjangoJSONEncoder
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from subprocess import Popen
 
 from mercury.models import AGEvent
-from mercury.serializers import AGMeasurementSerializer
 
 
 def serial_ports():
@@ -70,7 +69,7 @@ def call_script(uuid, port, fake):
         data = json.dumps(data, cls=DjangoJSONEncoder)
         command = command + "'" + data + "'"
         print(command)
-    subprocess.call(command, shell=True)
+    Popen(command, shell=True)
 
 
 def check_port(ser):
@@ -165,50 +164,3 @@ class RadioReceiverView(APIView):
         res["timeout"] = ser.timeout
 
         return Response(json.dumps(res), status=status.HTTP_200_OK)
-
-    def post(self, request, event_uuid=None):
-        """
-        The post receives sensor data through internet
-        Url example:
-        http://localhost:8000/radioreceiver/d81cac8d-26e1-4983-a942-1922e54a943d
-        Post Json Data Example
-        {
-          "sensor_id": 1,
-          "values": {
-            "power" : "1",
-            "speed" : "2",
-          }
-          "date" : 2020-03-11T20:20+01:00
-        }
-        """
-        # First check event_uuid exists
-        try:
-            event = AGEvent.objects.get(event_uuid=event_uuid)
-        except AGEvent.DoesNotExist:
-            event = False
-        if event is False:
-            return Response("Wrong uuid in url", status=status.HTTP_400_BAD_REQUEST)
-
-        json_data = request.data
-        if isinstance(json_data, str):
-            json_data = json.loads(json_data)
-        res = {"measurement_event": event_uuid}
-        dic = {
-            "measurement_timestamp": "date",
-            "measurement_sensor": "sensor_id",
-            "measurement_value": "values",
-        }
-
-        for d in dic:
-            if json_data.get(dic[d]) is None:
-                return Response(
-                    "Missing required params " + dic[d],
-                    status=status.HTTP_400_BAD_REQUEST,
-                )
-            res[d] = json_data[dic[d]]
-
-        serializer = AGMeasurementSerializer(data=res)
-        if serializer.is_valid():
-            serializer.save()
-            return Response("Saved Successfully", status=status.HTTP_200_OK)
-        return Response("The model fails to save", status=status.HTTP_400_BAD_REQUEST)
